Adding option to set empty strings null
Looks for empty strings in the arg and sets to None. This will cause the results to be treated like a null value when dumped via json.dumps. This is how data coming from a database looks. This is useful for targets like target-snowflake values can be empty i.e. null in the database rather than an empty string.